Follow good practices for stats page
